change: export ToPageBytes from file module to types module
