+++ src/file/page.rs
@@ -3,6 +3,8 @@ use core::fmt;
 use itertools::izip;
 use std::mem;
 
+use crate::types::page_bytes::ToPageBytes;
+
 #[derive(Debug)]
 enum PageError {
 	BufferSizeExceeded,
@@ -14,27 +16,6 @@ impl fmt::Display for PageError {
 		match self {
 			&PageError::BufferSizeExceeded => write!(f, "buffer size exceeded"),
 		}
-	}
-}
-
-pub trait ToPageBytes {
-	fn to_page_bytes(&self) -> Vec<u8>;
-}
-impl ToPageBytes for i32 {
-	fn to_page_bytes(&self) -> Vec<u8> {
-		self.to_be_bytes().to_vec()
-	}
-}
-impl ToPageBytes for &[u8] {
-	fn to_page_bytes(&self) -> Vec<u8> {
-		let mut v = (self.len() as i32).to_page_bytes();
-		v.append(&mut self.to_vec());
-		v
-	}
-}
-impl ToPageBytes for String {
-	fn to_page_bytes(&self) -> Vec<u8> {
-		self.as_bytes().to_page_bytes()
 	}
 }
